Remove EmailUtilities
